Replace old compression module with new one
Remove the old `compression` module and rename the new `ans` to
`compression`.
