+++ compressed_dynamic_word_embeddings/src/lib.rs
@@ -1,4 +1,3 @@
-pub mod ans;
 pub mod compression;
 pub mod embedding_file;
 pub mod random_access_reader;
